Support boolean props which use "0"/"1"
There are too many properties which treats "0" and "1" as its boolean
value, not "false" and "true". By setting integer_as_bool as true, one
can make setter use "0" and "1" instead. Note that this flag only works
with Boolean or BooleanList, and will emit an error if used with other
types.

Bug: 126261296
Test: 1) see compiling with various properties succeeds.
Test: 2) try to call setter with and without integer_as_bool.
Change-Id: Idf51058d0ad34312784ff28feb6a76666817b836