Always use en_US for string ops
Some languages, such as Turkish, do not handle casing well. So,
when this is set as the default language, any code that uses
toLowerCase() or toUpperCase() will get potentially incorrect
values.

In this case, we only do these operations on internal names
so we can safely use the en_US locale for all case operations.

Bug: 132320580
Test: Manual. Change locale to tr_TR and device reboots
Change-Id: Idb28452cb035e27d04a3fe52adbf93d88043334c