Allow prop removal if legacy_prop_name exists
Even if a prop is not deprecated, it can be removed in the next release
if a legacy_prop_name is identical to its prop_name. The following
example shows one of possible scenarioes.

// latest.txt
prop {
    api_name: "legacy_prop"
    prop_name: "legacy.prop.name"
}

// current.txt
prop {
    api_name: "new_prop"
    prop_name: "new.prop.name"
    legacy_prop_name: "legacy.prop.name"
}

"legacy_prop" is removed in the current version, but "new_prop" has
a legacy_prop_name which is identical to prop_name of "legacy_prop", so
removal of "legacy_prop" is allowed.

Bug: 152824209
Test: sysprop_test
Change-Id: I8e4b24409ab2113513b2dd992f245e5a2d4787a4